Factored out a getExpressionAliases method, moved alias methods into their own file
