+++ sql/analyzer/assign_indexes.go
@@ -38,42 +38,17 @@ func assignIndexes(ctx *sql.Context, a *Analyzer, node sql.Node) (map[string]*in
 		}
 	}()
 
-	aliases := make(map[string]sql.Expression)
-	var (
-		err error
-		fn  func(node sql.Node) bool
-	)
-	fn = func(n sql.Node) bool {
-		if n == nil {
-			return true
-		}
+	exprAliases := getExpressionAliases(node)
 
-		if prj, ok := n.(*plan.Project); ok {
-			for _, ex := range prj.Expressions() {
-				if alias, ok := ex.(*expression.Alias); ok {
-					if _, ok := aliases[alias.Name()]; !ok {
-						aliases[alias.Name()] = alias.Child
-					}
-				}
-			}
-		} else {
-			for _, ch := range n.Children() {
-				plan.Inspect(ch, fn)
-			}
-		}
-
-		return true
-	}
-
+	var err error
 	plan.Inspect(node, func(node sql.Node) bool {
 		filter, ok := node.(*plan.Filter)
 		if !ok {
 			return true
 		}
-		fn(filter.Child)
 
 		var result map[string]*indexLookup
-		result, err = getIndexes(ctx, filter.Expression, aliases, a)
+		result, err = getIndexes(ctx, filter.Expression, exprAliases, a)
 		if err != nil {
 			return false
 		}
